Implement direct HD44780 display

+++ gpio4k/src/commonMain/kotlin/dev/thechilli/gpio4k/CharacterDisplay.kt
@@ -74,5 +74,5 @@ interface CharacterDisplay {
     /**
      * Reads the current address.
      */
-    fun readAddress(): Int
+    fun readAddress(): UByte
 }
